+++ equinox-web/Domain/Aggregate.fs
@@ -8,6 +8,7 @@ module Events =
         | Happened
         | Compacted of Compacted
         interface TypeShape.UnionContract.IUnionContract
+    let codec = FsCodec.NewtonsoftJson.Codec.Create<Event>()
 
 module Folds =